ec2_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.32.4
Flask-SocketIO==5.3.6
eventlet==0.36.1
gunicorn==21.2.0
boto3==1.34.162
google-auth==2.23.4
google-auth-oauthlib==1.2.0